refactor: split get_estimation_guide into focused files
